Improve thirdparty creation (#29801)
#29801
A large number of users need to create third parties that are not companies and associate a contact simultaneously (construction or service company). When using the hidden option "THIRDPARTY_SUGGEST_ALSO_ADDRESS_CREATION" this modification allows:

a copy of the mobile number provided on the contact form,
to hide unnecessary fields for creating individuals.